Add a method to MGTwitterEngine which lets us update our profile. Add the respective profile view to the account edit view, and save the values when saving the account.
The '''major''' problem with this is that Twitter lies to us: when it updates the values, it doesn't update the values when responding to us. We never are told when new values take hold, so the next time the view is loaded it has the old values.

--HG--
extra : convert_revision : svn%3A32148f9c-d742-0410-bc91-dd12f8e4a632/trunk%4026133