This logic was driving me crazy. Remove an {{{if()}}} whose functionality in the outlineView validateDrop was questionable at best:  * If we're not dropping on an object  * If we can't sort manually ->  * Check the index for inserting.  * If the index is the same as its current, return that we can't move.  * Else return that we can move to the new index.
Well, shortly below that is the same exact logic, except after a check for a {{{nil}}} {{{item}}}, which was safeguarding dropping on certain parts of the contact list where a drop would occur on the contact list itself.

Adds logic to prevent a drag if we're:
 * Not able to manually sort
 * Dragging into an item which already contains the object

There's still an outstanding bug:
 * Before, we'd try and calculate the new index based on total contained objects.
 * When calculating the new index, we'd exceed the _visual_ object bounds.
I've half-fixed this by making inserts into groups go to index 0. This needs to be fixed.

--HG--
extra : convert_revision : svn%3A32148f9c-d742-0410-bc91-dd12f8e4a632/trunk%4026294