Change AdiumURLHandling into an AIPlugin subclass, which allows us to create an advanced preference pane to control its actions. Adds a few useful features:  * The advanced preference pane lets the user pick on a service-by-service level the default application.  * Added an option to enforce Adium being the default for all of its services.  * Adds a button to set Adium to the default for all of our services. This changes the way AIAdium calls the URL handler, to a notification (from a direct class method) which also allows the WKMV to post the notification when a URL belonging to us is clicked. With this change, running multiple copies of Adium is much easier. :)
The old "first run URL registration" preference is still respected, so we should only register ourself on the initial pass. We don't bother the user afterwards, they can come back and set us how they want later on.

The reason this is not in General preferences is because no single service or application can be listed. Does iChat support all of our IM services? What does setting it to the default application do for schemes it doesn't support? It's too hard of a concept to display, without a "Default IM: (Configure)" which is tacky, and this basically is the same thing anyway, just stuffed into a more clear area.

Fixes #158.

--HG--
extra : convert_revision : svn%3A32148f9c-d742-0410-bc91-dd12f8e4a632/trunk%4026704