Tentative fix for #11813. References #11813.
There are two issues here; one is causing 11813, the other is just something I found while reading over the code.

First issue: -prefs contains an unsynchronized increment of (*myUsersOfGlobalPreferences). This lead to a race condition with clearing the prefs cache, leading to 11813.
Second issue: @synchronized(nil) does nothing (see https://twitter.com/ddribin/status/1354909749). Because of cache clearing, *myGlobalPrefs can be nil. This changes it so that we empty the dictionary instead of releasing it and setting the pointer to nil, and initializes the dictionary the first time any AIPreferenceContainer is -init'd.

Conclusion: The threading/locking in AIPreferenceContainer is overly complex and error-prone. We should investigate a simpler replacement (I have some ideas, but they're NDA'd right now).