Avoid creating a bad account on the wrong service for "preferred account" guesses.
The preferred account for a meta contact was attempting to create contacts on its preferred account, even if the requested contact was on a different service. This know makes sure we're returning the correct account type, and adds an assert to make diagnosing this issue easier in the future. Fixes #11866.

--HG--
extra : rebase_source : 2cdfdbc8436b7240610bd20281165e970ebf46f7