An image uploader plugin. Found in the same locations as "Replace with Shortened URL."
Currently the only supported service is pic.im, and even then not fully since we don't have an API key for uploading yet. Once we get one it will be simple to finish this plugin off.

I'm not willing to add any services that require a username/password, so most of the traditional services will never be available.