Fixed retaining delegate property, introduced in 1d5619b64c61.
Believe it or not, this caused an *under*-release crash when a FT finished with the progress window closed: The delegate is a progress row, which is normally owned solely by the progress window controller. With this retention keeping the row alive, Adium crashed when the row went to talk to the now-dead progress window controller.

--HG--
extra : rebase_source : f39e692e1b585751a97cba9240a5eb51ad8a229b