Set the Contact Info Inspector to not "hide on deactivate." Fixes #12714.
The inspector contains a fair amount of useful information, especially in the case of Facebook contacts or Address Book contacts. It's not as much of an inspector as other applications deal with: it's not just modifying the information it's inspecting, it's also presenting more information.

A future version of Adium could go the route of dividing a "get info" and "show inspector" version of this, but I've run into this situation enough that it drives me crazy having to tab back and forth for more information.
(transplanted from 67cf100919f4d0ba69117ad0ac3eb62189efb666)

--HG--
extra : rebase_source : 0168b5f6c935dc99f5e02fce31f34fc366af405f