Patch from Jeff Mahoney to fix a crash when closing the Novell Groupwise account options. Fixes #16041.
"While trying to change the port number for my Novell Groupwise account, Adium crashed consistently when the account options window was closed.

I tracked it down to a double release of view_setup which, at least for Groupwise, is the same as view_options. Releasing them separately caused a problem later with the autorelease was later processed.

This patch only frees both if they are different values."

--HG--
branch : adium-1.5.2